requirements.txt
pandas
python-dotenv
streamlit
streamlit-cookies-controller
redis